refactor: consolidate CartContext into CartProvider for improved context management
